src/main/java/com/example/filmspringboot/FilmService.java
package com.example.filmspringboot;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import java.util.List;
import java.util.Optional;
@Service
public class FilmService {

    private final FilmRepository filmRepository;

    @Autowired
    public FilmService(FilmRepository filmRepository) {
        this.filmRepository = filmRepository;
    }

    public List<Film> getAllFilms() {
        return filmRepository.findAll();
    }

    public Optional<Film> getFilmById(Long id) {
        return filmRepository.findById(id);
    }

    public void addFilm(Film film) {
        filmRepository.save(film);
    }

    public void updateFilm(Film film) {
        filmRepository.save(film);
    }

    public void deleteFilm(Long id) {
        filmRepository.deleteById(id);
    }

    public List<Film> findFilmsByLastName(String lastName) {
        return filmRepository.findByLastNameIgnoreCase(lastName);
    }
}